+++ lib/core/storage/hive/default_hive_box_names.dart
@@ -1,4 +1,5 @@
-const kApiCacheBoxName = "kickin_api_cache_box";
+const kApiCacheBoxName = "kickin_api_cache";
 const kAppBoxName = "kickin_box";
 const kSecureBoxName = "kickin_secure_box";
 const kRiverpodCacheBoxName = "kickin_riverpod_cache_box";
+const kKickinCache = "kickin_cache";